Revert configuration to match AEP Gradle Plugin v3.0.2
Although we explicitly set Kotlin compatibility to 1.5 in our modules,
this does not extend to transitive dependencies, leading to issues—particularly
with Jetpack Compose dependencies, which are only compatible with Kotlin 1.9 with
the current plgin version 3.1.0.

AEP Extensions use AEP GP 3.0.2 so this should not affect the current production versions.
Moving forward, we will aim to maintain Kotlin compatibility with version 1.8,
even as we upgrade AGP and Gradle.